20 06

+++ src/controller/tag/DeleteTag.java
@@ -2,6 +2,7 @@ package controller.tag;
 
 import java.io.IOException;
 import java.sql.SQLException;
+import java.util.regex.Pattern;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -11,7 +12,10 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import org.apache.commons.text.StringEscapeUtils;
+
 import dao.TagDao;
+import model.Tag;
 import model.User;
 
 @WebServlet("/deleteTag")
@@ -47,31 +51,34 @@ public class DeleteTag extends HttpServlet {
 		
 		User user = (User) session.getAttribute("user");
 		if(user!=null) {
-			String from = request.getParameter("from").trim();
-			
-			System.out.println("From: ");
-			System.out.println(from);
-			
-			int id = Integer.parseInt(request.getParameter("id"));
-			
-			tagDao.deleteTag(id);
-			
-			if (from.equals("dashboard")) {
-				response.sendRedirect("listDashboard");
-			} else if (from.equals("tododay")) {
-				response.sendRedirect("listTodo");
-			} else if (from.equals("todoweek")) {
-				response.sendRedirect("listTodoThisWeek");
+			if(request.getParameter("from").equals("dashboard") || request.getParameter("from").equals("tododay") || request.getParameter("from").equals("todoweek") || request.getParameter("from").equals("todomonth")) {
+				
+				String from = request.getParameter("from");
+				
+				int id = Integer.parseInt(request.getParameter("id"));
+				
+				tagDao.deleteTag(id);
+				
+				if (from.equals("dashboard")) {
+					response.sendRedirect("listDashboard");
+				} else if (from.equals("tododay")) {
+					response.sendRedirect("listTodo");
+				} else if (from.equals("todoweek")) {
+					response.sendRedirect("listTodoThisWeek");
+				} else {
+					response.sendRedirect("listTodoThisMonth");
+				}	
+				
 			} else {
-				response.sendRedirect("listTodoThisMonth");
+				RequestDispatcher dispatcher;
+				dispatcher = request.getRequestDispatcher("error.jsp");
+				dispatcher.forward(request, response);
 			}
-		}else {
+			
+		} else {
 			System.out.println("Nguoi dung null");
-			
 			RequestDispatcher dispatcher;
-			
 			dispatcher = request.getRequestDispatcher("index.jsp");
-			
 			dispatcher.forward(request, response);
 		}
 	}
